+++ subtitle2go.py
@@ -15,328 +15,23 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
-import os
-
-# Load Kaldi
-from kaldi.asr import NnetLatticeFasterRecognizer, LatticeLmRescorer, LatticeRnnlmPrunedRescorer
-from kaldi.rnnlm import RnnlmComputeStateComputationOptions
-from kaldi.decoder import LatticeFasterDecoderOptions
-from kaldi.lat.functions import ComposeLatticePrunedOptions
-from kaldi.lat.align import read_lexicon_for_word_align
-from kaldi.fstext import SymbolTable, shortestpath, indices_to_symbols
-from kaldi.fstext.utils import get_linear_symbol_sequence
-from kaldi.nnet3 import NnetSimpleComputationOptions
-from kaldi.util.table import SequentialMatrixReader
-from kaldi.lat import functions
-from kaldi.transform import cmvn
-
 # Audio Segmentation
-from simple_endpointing import process_wav
-
+from kaldi_decoder import asr, kaldi_time_to_seconds
 # Interpunctuation
-from rpunct import RestorePuncts
 
 import yaml
 import argparse
-import ffmpeg
 import segment_text
 import slide_stripper
-import json
 import time
 import sys
 
-import requests
-
-
-
-start_time = time.time()
-
-kaldi_feature_factor = 3.00151874884282680911
-
-class output_status():
-    def __init__(self, filename, fn_short_hash, redis=False):
-        if redis:
-            try:
-                import redis
-                self.red = redis.StrictRedis(charset='utf-8', decode_responses=True)
-            except ImportError:
-                print('Redis is not available. Disabling redis option.')
-                redis = False
-
-            self.redis_server_channel = 'subtitle2go'
-        self.redis = redis
-
-        self.filename = filename
-        self.fn_short_hash = fn_short_hash
-
-    def publish_status(self, status):
-        print(f'{filename=} {self.fn_short_hash=} {status=}')
-        if self.redis:
-            self.red.publish(self.redis_server_channel, json.dumps({'pid': os.getpid(), 'time': time.time(), 'start_time': start_time,
-                                                    'file_id': self.fn_short_hash, 'filename': self.filename,
-                                                    'status': status}))
-
-# Make sure a fpath directory exists
-def ensure_dir(fpath):
-    directory = os.path.dirname(fpath)
-    if not os.path.exists(directory):
-        os.makedirs(directory)
-
-
-def preprocess_audio(filename, wav_filename):
-    # Use ffmpeg to convert the input media file (any format!) to 16 kHz wav mono
-    (
-        ffmpeg
-            .input(filename)
-            .output(wav_filename, acodec='pcm_s16le', ac=1, ar='16k')
-            .overwrite_output()
-            .run(quiet=True)
-    )
-
-def recognizer(decoder_yaml_opts, models_dir):
-    decoder_opts = LatticeFasterDecoderOptions()
-    decoder_opts.beam = decoder_yaml_opts['beam']
-    decoder_opts.max_active = decoder_yaml_opts['max-active']
-    decoder_opts.lattice_beam = decoder_yaml_opts['lattice-beam']
-
-    decodable_opts = NnetSimpleComputationOptions()
-    decodable_opts.acoustic_scale = decoder_yaml_opts['acoustic-scale']
-    decodable_opts.frame_subsampling_factor = 3 # decoder_yaml_opts['frame-subsampling-factor'] # 3
-    decodable_opts.frames_per_chunk = 150
-    fr = NnetLatticeFasterRecognizer.from_files(
-        models_dir + decoder_yaml_opts['model'],
-        models_dir + decoder_yaml_opts['fst'],
-        models_dir + decoder_yaml_opts['word-syms'],
-        decoder_opts=decoder_opts, decodable_opts=decodable_opts)
-
-    return fr
-
-# This method contains all Kaldi related calls and methods
-def Kaldi(config_file, scp_filename, spk2utt_filename, segments_filename, do_rnn_rescore, segments_timing, lm_scale, acoustic_scale):
-
-    models_dir = 'models/'
-
-    # Read yaml File
-    with open(config_file, 'r') as stream:
-        model_yaml = yaml.safe_load(stream)
-    decoder_yaml_opts = model_yaml['decoder']
-
-    # Construct recognizer
-    fr = recognizer(decoder_yaml_opts, models_dir)
-
-    # Check if cmvn is set
-    cmvn_transformer = None
-    if decoder_yaml_opts.get('global-cmvn-stats'):
-        cmvn_transformer = cmvn.Cmvn(40)
-        cmvn_transformer.read_stats(f'{models_dir}{decoder_yaml_opts["global-cmvn-stats"]}')
-
-    # Construct symbol table
-    symbols = SymbolTable.read_text(models_dir + decoder_yaml_opts['word-syms'])
-    
-    # Define feature pipelines as Kaldi rspecifiers
-    feats_rspec = (f'ark:extract-segments scp,p:{scp_filename} {segments_filename} ark:- | compute-mfcc-feats --config={models_dir}{decoder_yaml_opts["mfcc-config"]} ark:- ark:- |')
-    
-    ivectors_rspec = (
-            (f'ark:extract-segments scp,p:{scp_filename} {segments_filename} ark:- | compute-mfcc-feats --config={models_dir}{decoder_yaml_opts["mfcc-config"]} '
-            f'ark:- ark:- | '
-            f'ivector-extract-online2 --config={models_dir}{decoder_yaml_opts["ivector-extraction-config"]} '
-            f'ark:{spk2utt_filename} ark:- ark:- |'))
-    
-    
-    rnn_rescore_available = 'rnnlm' in decoder_yaml_opts
-
-    if do_rnn_rescore and not rnn_rescore_available:
-        status.publish_status("Warning, disabling RNNLM rescoring since 'rnnlm' is not in the decoder options of the .yaml config.")
-
-    if do_rnn_rescore and rnn_rescore_available:
-        status.publish_status('Loading language model rescorer.')
-        rnn_lm_folder = models_dir + decoder_yaml_opts['rnnlm'] 
-        arpa_G = models_dir + decoder_yaml_opts['arpa'] 
-        old_lm = models_dir + decoder_yaml_opts['fst'] 
-
-        print(f'Loading RNNLM rescorer from:{rnn_lm_folder} with ARPA from:{arpa_G} FST:{old_lm}')
-        # Construct RNNLM rescorer
-        symbols = SymbolTable.read_text(rnn_lm_folder+'/config/words.txt')
-        rnnlm_opts = RnnlmComputeStateComputationOptions()
-        rnnlm_opts.bos_index = symbols.find_index('<s>')
-        rnnlm_opts.eos_index = symbols.find_index('</s>')
-        rnnlm_opts.brk_index = symbols.find_index('<brk>')
-        compose_opts = ComposeLatticePrunedOptions()
-        compose_opts.lattice_compose_beam = 6
-        print(f'rnnlm-get-word-embedding {rnn_lm_folder}/word_feats.txt {rnn_lm_folder}/feat_embedding.final.mat -|')
-        print(f'{rnn_lm_folder}/final.raw')
-        rescorer = LatticeRnnlmPrunedRescorer.from_files(
-            arpa_G,
-            f'rnnlm-get-word-embedding {rnn_lm_folder}/word_feats.txt {rnn_lm_folder}/feat_embedding.final.mat -|',
-            f'{rnn_lm_folder}/final.raw', lm_scale=lm_scale, acoustic_scale=acoustic_scale, max_ngram_order=4,
-            use_const_arpa=True, opts=rnnlm_opts, compose_opts=compose_opts)
-
-    did_decode = False
-    decoding_results = []
-
-    segmentcounter = 1
-    with SequentialMatrixReader(feats_rspec) as f, \
-            SequentialMatrixReader(ivectors_rspec) as i:
-            for (fkey, feats), (ikey, ivectors) in zip(f, i):
-                status.publish_status(f'Decoding segment {segmentcounter} of {len(segments_timing)}.')
-                if cmvn_transformer:
-                    cmvn_transformer.apply(feats)
-                did_decode = True
-                assert (fkey == ikey)
-                out = fr.decode((feats, ivectors))
-                if do_rnn_rescore:
-                    lat = rescorer.rescore(out['lattice'])
-                else:
-                    lat = out['lattice']
-                best_path = functions.compact_lattice_shortest_path(lat)
-                words, _, _ = get_linear_symbol_sequence(shortestpath(best_path))
-                timing = functions.compact_lattice_to_word_alignment(best_path)
-                decoding_results.append((words, timing))
-                segmentcounter+=1
-
-    # Concatenating the results of the segments and adding an offset to the segments
-    words = []
-    timing = [[],[],[]]
-    for result in decoding_results:
-        words.extend(result[0])
-
-    for result, offset in zip(decoding_results, segments_timing):
-        if result[1][1]:
-            timing[0].extend(result[1][0])
-            # start = map(lambda x: int(x + (offset[0] / kaldi_feature_factor)), result[1][1])
-            start = [x + (offset[0] / kaldi_feature_factor) for x in result[1][1]]
-            kaldi_time_to_seconds(offset[0] / kaldi_feature_factor, seperator=".")
-            kaldi_time_to_seconds(start[0], seperator=".")
-
-            timing[1].extend(start)
-            timing[2].extend(result[1][2])
-    starting = 0
-    temp_timing = [[], [], []]
-
-    # Maps words to the numbers
-    words = indices_to_symbols(symbols, timing[0])
-
-    # Creates the datastructure (Word, begin(Frames), end(Frames))
-    assert(len(words) == len(timing[1]))
-    assert(len(timing[1]) == len(timing[2]))
-    vtt = list(map(list, zip(words, timing[1], timing[2])))
-
-    if debug_word_timing:
-        with open('debug_output.txt', 'w') as f:
-            for element in vtt:
-                f.write(f'{element[1]} {kaldi_time_to_seconds(element[1], ".")} {kaldi_time_to_seconds(element[1] + element[2], ".")} {element[2]} {element[0]}\n')
-
-    return vtt, did_decode, words
-
-# This is the asr function that converts the videofile, split the video into segments and decodes
-def asr(filenameS_hash, filename, asr_beamsize=13, asr_max_active=8000, acoustic_scale=1.0, lm_scale=0.5,
-         do_rnn_rescore=False, config_file='models/kaldi_tuda_de_nnet3_chain2_de_722k.yaml'):
-
-    print(f"{filenameS_hash=}")
-
-    scp_filename = f'tmp/{filenameS_hash}.scp'
-    segments_filename = f'tmp/{filenameS_hash}_segments'
-    wav_filename = f'tmp/{filenameS_hash}.wav'
-    spk2utt_filename = f'tmp/{filenameS_hash}_spk2utt'
-
-    # Audio extraction
-    status.publish_status('Extract audio.')
-    
-    try:
-        preprocess_audio(filename, wav_filename)
-    except ffmpeg.Error as e:
-        status.publish_status('Audio extraction failed.')
-        status.publish_status(f'Error message is: {e.stderr}')
-        send_error()
-        sys.exit(-1)
-
-    status.publish_status('Audio extracted.')
-
-    # Segmentation
-    status.publish_status('Audio segmentation.')
-
-    try:
-        segments_filenames, segments_timing = process_wav(wav_filename)
-    except Exception as e:
-        status.publish_status('Audio segmentation failed.')
-        status.publish_status(f'Error message is: {e}')
-        send_error()
-        sys.exit(-1)
-    
-    # Write scp and spk2utt file
-    with open(scp_filename, 'w') as wavscp, open(spk2utt_filename, 'w') as spk2utt:
-        # segmentFilename = wav_filename.rpartition('.')[0]
-        wavscp.write(f'{filenameS_hash} {wav_filename}\n')
-
-        for i in range(len(segments_timing)):
-            count_str = "%.4d" % i    
-            spk2utt.write(f'{filenameS_hash} {filenameS_hash}_{count_str}\n')
-
-    # Decode wav files
-    status.publish_status('Start ASR.')
-    vtt, did_decode, words = Kaldi(config_file, scp_filename, spk2utt_filename, segments_filename, do_rnn_rescore, segments_timing, lm_scale, acoustic_scale)
-    if did_decode:
-        status.publish_status('ASR finished.')
-    else:
-        status.publish_status('ASR error.')
-        send_error()
-        sys.exit(-1)
-
-    # Cleanup tmp files
-    try:
-        os.remove(scp_filename)
-        os.remove(wav_filename)
-        os.remove(spk2utt_filename)
-        os.remove(segments_filename)
-        status.publish_status(f'Removed temporary files: {scp_filename=}, {wav_filename=}, {spk2utt_filename=}, {segments_filename=}')
-    except Exception as e:
-        status.publish_status(f'Removing files failed.')
-        status.publish_status(f'Error message is: {e}')
-        send_error()
-
-    status.publish_status('VTT finished.')
-
-    return vtt, words
-
-
-# Adds interpunctuation to the Kaldi output
-def interpunctuation(vtt, words, filenameS_hash, model_punctuation, uppercase):
-
-    status.publish_status('Starting interpunctuation.')
-
-    # BERT
-    text = str(' '.join(words))
-    if uppercase:
-        text = text.lower()
-    rpunct = RestorePuncts(model=model_punctuation)
-
-    punct = rpunct.punctuate(text)
-    punct = punct.replace('.', '. ').replace(',', ', ').replace('!', '! ').replace('?', '? ')
-    punct = punct.replace('  ', ' ')
-    punct_list = punct.split(' ')
-
-    # punct_list = file_punct.read().split(' ')
-    vtt_punc = []
-    for a, b in zip(punct_list, vtt):  # Replaces the adapted words with the (capitalization, period, comma) with the new ones
-        vtt_punc.append([a, b[1], b[2]])
-
-    status.publish_status('Adding interpunctuation finished.')
-
-    return vtt_punc
-
-
-def kaldi_time_to_seconds(value, seperator):
-    time = value * kaldi_feature_factor / 100
-    time_start =    (f'{int(time / 3600):02}:'
-                            f'{int(time / 60 % 60):02}:'
-                            f'{int(time % 60):02}'
-                            f'{seperator}'
-                            f'{int(time * 1000 % 1000):03}')
-    return time_start
+from punctuation import interpunctuation
+from utils import output_status, ensure_dir
 
 # This creates a segmentation for the subtitles and make sure it can still be mapped to the Kaldi tokenisation
-def segmentation(vtt, model_spacy, beam_size, ideal_token_len, len_reward_factor, comma_end_reward_factor,
-                 sentence_end_reward_factor):
+def vtt_segmentation(vtt, model_spacy, beam_size, ideal_token_len, len_reward_factor, comma_end_reward_factor,
+                     sentence_end_reward_factor, status):
     sequences = []
 
     status.publish_status('Start text segmentation.')
@@ -393,7 +88,7 @@ def segmentation(vtt, model_spacy, beam_size, ideal_token_len, len_reward_factor
     return sequences
 
 # Creates the subtitle in the desired subtitleFormat and writes to filenameS (filename stripped) + subtitle suffix
-def create_subtitle(sequences, subtitle_format, filenameS):
+def create_kaldi_subtitle(sequences, subtitle_format, filenameS, status):
     status.publish_status('Start creating subtitle.')
 
     try:
@@ -421,20 +116,21 @@ def create_subtitle(sequences, subtitle_format, filenameS):
     except Exception as e:
         status.publish_status('Subtitle creation failed.')
         status.publish_status(f'error message is: {e}')
-        send_error()
+        status.send_error()
         sys.exit(-1)
 
     status.publish_status('Finished subtitle creation.')
 
-def send_error():
-    if (callback_url):
-        d = {'message': 'false'}
-        r = requests.put(callback_url, data = d)
-
-def send_success():
-    if (callback_url):
-        d = {'message': 'true'}
-        r = requests.put(callback_url, data = d)
+def pykaldi_subtitle(status):
+    vtt, words = asr(filenameS_hash, filename=filename, asr_beamsize=args.asr_beam_size,
+                     asr_max_active=args.asr_max_active, acoustic_scale=args.acoustic_scale,
+                     do_rnn_rescore=args.rnn_rescore, config_file=model_kaldi, status=status)
+    vtt = interpunctuation(vtt, words, filenameS_hash, model_punctuation, uppercase, status=status)
+    sequences = vtt_segmentation(vtt, model_spacy, beam_size=args.segment_beam_size, ideal_token_len=args.ideal_token_len,
+                                 len_reward_factor=args.len_reward_factor,
+                                 sentence_end_reward_factor=args.sentence_end_reward_factor,
+                                 comma_end_reward_factor=args.comma_end_reward_factor, status=status)
+    create_kaldi_subtitle(sequences, subtitle_format, filenameS, status=status)
 
 
 if __name__ == '__main__':
@@ -502,12 +198,13 @@ if __name__ == '__main__':
     # Positional argument, without (- and --)
     parser.add_argument('filename', help='The path of the mediafile', type=str)
 
+
+
     args = parser.parse_args()
     filenameS = args.filename.rpartition('.')[0] # Filename without file extension
     filename = args.filename
     subtitle_format = args.subtitle
     pdf_path = args.pdf
-    model_kaldi = args.model_yaml
     debug_word_timing = args.debug
     file_id=args.id
     callback_url=args.callback_url
@@ -520,7 +217,7 @@ if __name__ == '__main__':
     ensure_dir('tmp/')
 
     # Init status class
-    status = output_status(redis=args.with_redis_updates, filename=filename, fn_short_hash=filenameS_hash)
+    status = output_status(redis=args.with_redis_updates, filename=filename, fn_short_hash=filenameS_hash, callback_url=callback_url)
 
     # Language selection
     language = args.language
@@ -539,18 +236,7 @@ if __name__ == '__main__':
     if (pdf_path):
         slides = slide_stripper.convert_pdf(pdf_path)
 
-    vtt, words = asr(filenameS_hash, filename=filename, asr_beamsize=args.asr_beam_size,
-                     asr_max_active=args.asr_max_active, acoustic_scale=args.acoustic_scale,
-                     do_rnn_rescore=args.rnn_rescore, config_file=model_kaldi)
-    
-    vtt = interpunctuation(vtt, words, filenameS_hash, model_punctuation, uppercase)
-
-    sequences = segmentation(vtt, model_spacy, beam_size=args.segment_beam_size, ideal_token_len=args.ideal_token_len,
-                             len_reward_factor=args.len_reward_factor,
-                             sentence_end_reward_factor=args.sentence_end_reward_factor,
-                             comma_end_reward_factor=args.comma_end_reward_factor)
-
-    create_subtitle(sequences, subtitle_format, filenameS)
+    pykaldi_subtitle(status)
 
     status.publish_status('Job finished successfully.')
-    send_success()
+    status.send_success()
